refactor(meetings): split platform adapter contract
