+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -82,7 +82,7 @@ a5f59c9acbcaa3f82247bf806eb5ba08032373fb853719f0ec9457690f16fc70  module/media-m
 7a5a1de743fff9d5a86515b501ffd88e7fa49970c9f46a5153816e2ebaef7bb0  module/media-store
 c5e3eb1a584f4b8126d9d6c177a840ec9103671e8d1242634ee67db9b5b9e573  module/media-understanding
 c0ffaed532578cf33493992e1ff806b2268b8e3774a92edbaede5cf5bda162a6  module/media-understanding-runtime
-2430612f0f51bb2c581284ac5f65f9968009f126da85c9bee1fda14b215a0618  module/meeting-runtime
+8e318232d62a5a75e5c39e38d924083efae2ff6f9a7dea25cabd5b84593083f8  module/meeting-runtime
 d09ddb38c9d20a41fcf4cedaf6a2c28eb5333a9fba2287e34b458fa5caba751c  module/memory-core-host-engine-foundation
 646773d8282a2ac6a89101685c406200935ddb3459a2830e820da132fa433b3c  module/memory-host-core
 87b7a3206346c0d4b294fb3a2395cbaabc3e73ff8b1b9ea925bc3aade3e52687  module/messaging-targets
